requirements.txt
asgiref==3.8.1
autopep8==2.1.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-silk==5.0.3
django-timescaledb==0.2.13
djangorestframework==3.15.1
drf-nested-routers==0.93.5
future==1.0.0
gprof2dot==2022.7.29
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycodestyle==2.11.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
routers==0.10.1
serializers==0.2.4
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
zstandard==0.22.0
